+++ tests/Dfe.EarlyYearsQualification.E2ETests/tests/journey/deny-public-access.spec.ts
@@ -2,7 +2,7 @@
 import {
     checkUrl,
     checkValue
-} from '../shared/processLogic';
+} from '../shared/playwrightWrapper';
 
 test.describe('A spec used to check a new user is challenged to enter the secret', () => {
     test("should redirect the user to the challenge page", async ({page}) => {
