+++ tests/Dfe.EarlyYearsQualification.E2ETests/tests/journey/journey.spec.ts
@@ -18,7 +18,7 @@ import {
     confirmAdditonalRequirementsAnswers,
     checkDetailsPage, 
     checkEmptyValue
-} from '../shared/processLogic';
+} from '../shared/playwrightWrapper';
 
 test.describe('A spec used to test the various routes through the journey', () => {
     test.beforeEach(async ({page, context}) => {
