+++ tests/Dfe.EarlyYearsQualification.E2ETests/tests/pages/qualification-not-on-list.spec.ts
@@ -1,5 +1,5 @@
 ﻿import {test} from '@playwright/test';
-import {startJourney, checkText, setCookie, journeyCookieName} from '../shared/processLogic';
+import {startJourney, checkText, setCookie, journeyCookieName} from '../shared/playwrightWrapper';
 
 
 test.describe('A spec that tests the qualification not on list page', () => {
