+++ tests/Dfe.EarlyYearsQualification.E2ETests/tests/pages/question.spec.ts
@@ -1,5 +1,5 @@
 ﻿import {test, expect} from '@playwright/test';
-import {startJourney, checkText, checkUrl, setCookie, journeyCookieName} from '../shared/processLogic';
+import {startJourney, checkText, checkUrl, setCookie, journeyCookieName} from '../shared/playwrightWrapper';
 
 
 test.describe("A spec that tests question pages", () => {
